requirements.txt
psycopg2-binary

pandas==1.5.2
python_daemon==3.0.1
Requests==2.31.0
smbus2==0.4.2
spidev==3.6
